Minor cleanups
git-svn-id: svn://messdev.no-ip.org/mess@4259 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/audio/cgenie.c
@@ -9,7 +9,7 @@
 #include "includes/cgenie.h"
 #include "sound/ay8910.h"
 
-static int control_port;
+static UINT8 control_port;
 
 READ8_HANDLER( cgenie_sh_control_port_r )
 {
